Refactor a lot of things related to dagger
