+++ android/app/src/main/java/com/hedvig/app/MainApplication.kt
@@ -10,7 +10,6 @@ import android.support.v7.app.AppCompatDelegate
 import androidx.work.Configuration
 import androidx.work.WorkManager
 import androidx.work.WorkerFactory
-import androidx.work.WorkerParameters
 import com.RNFetchBlob.RNFetchBlobPackage
 import com.airbnb.android.react.lottie.LottiePackage
 import com.apollographql.apollo.ApolloClient
@@ -19,9 +18,9 @@ import com.facebook.react.ReactApplication
 import com.facebook.react.ReactNativeHost
 import com.facebook.react.shell.MainReactPackage
 import com.facebook.soloader.SoLoader
+import com.hedvig.app.di.DaggerApplicationComponent
 import com.hedvig.app.react.ActivityStarterReactPackage
-import com.hedvig.app.service.PushNotificationWorker
-import com.hedvig.app.util.react.AsyncStorageNative
+import com.hedvig.app.react.NativeRoutingPackage
 import com.horcrux.svg.SvgPackage
 import com.imagepicker.ImagePickerPackage
 import com.jakewharton.threetenabp.AndroidThreeTen
@@ -60,7 +59,7 @@ class MainApplication : Application(), ReactApplication, HasActivityInjector, Ha
     lateinit var apolloClient: ApolloClient
 
     @Inject
-    lateinit var asyncStorageNative: AsyncStorageNative
+    lateinit var workerFactory: WorkerFactory
 
     private val mReactNativeHost = object : ReactNativeHost(this) {
         override fun getUseDeveloperSupport() = BuildConfig.DEBUG
@@ -115,14 +114,6 @@ class MainApplication : Application(), ReactApplication, HasActivityInjector, Ha
 
         AppCompatDelegate.setCompatVectorFromResourcesEnabled(true)
 
-
-        val workerFactory = object : WorkerFactory() {
-            override fun createWorker(
-                appContext: Context,
-                workerClassName: String,
-                workerParameters: WorkerParameters
-            ) = PushNotificationWorker(appContext, workerParameters, apolloClient, asyncStorageNative)
-        }
         WorkManager
             .initialize(this, Configuration.Builder().setWorkerFactory(workerFactory).build())
     }
