--- android/app/src/main/java/com/hedvig/app/di/AppViewModelModule.kt
@@ -1,16 +1,0 @@
-package com.hedvig.app.di
-
-import android.arch.lifecycle.ViewModel
-import com.hedvig.app.feature.chat.ChatViewModel
-import dagger.Binds
-import dagger.Module
-import dagger.multibindings.IntoMap
-
-@Module
-abstract class AppViewModelModule {
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(ChatViewModel::class)
-    abstract fun chatViewModel(viewModel: ChatViewModel): ViewModel
-}
