+++ android/app/src/main/java/com/hedvig/app/di/viewmodel/ViewModelFactory.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.di
+package com.hedvig.app.di.viewmodel
 
 import android.arch.lifecycle.ViewModel
 import android.arch.lifecycle.ViewModelProvider
@@ -10,7 +10,7 @@ import javax.inject.Singleton
 class ViewModelFactory @Inject constructor(
     private val viewModels: MutableMap<Class<out ViewModel>, Provider<ViewModel>>
 ) :
-        ViewModelProvider.Factory {
+    ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel?> create(modelClass: Class<T>): T = viewModels[modelClass]?.get() as T
 }
