+++ android/app/src/main/java/com/hedvig/app/di/worker/ChildWorkerFactory.kt
@@ -1,0 +1,9 @@
+package com.hedvig.app.di.worker
+
+import android.content.Context
+import androidx.work.Worker
+import androidx.work.WorkerParameters
+
+interface ChildWorkerFactory {
+    fun create(context: Context, params: WorkerParameters): Worker
+}
