+++ android/app/src/main/java/com/hedvig/app/di/worker/WorkerKey.kt
@@ -1,0 +1,9 @@
+package com.hedvig.app.di.worker
+
+import androidx.work.Worker
+import dagger.MapKey
+import kotlin.reflect.KClass
+
+@Retention(AnnotationRetention.RUNTIME)
+@MapKey
+annotation class WorkerKey(val value: KClass<out Worker>)
