+++ android/app/src/main/java/com/hedvig/app/feature/claims/ui/pledge/HonestyPledgeBottomSheet.kt
@@ -8,7 +8,7 @@ import android.support.annotation.IdRes
 import android.view.LayoutInflater
 import androidx.navigation.findNavController
 import com.hedvig.app.R
-import com.hedvig.app.di.ViewModelFactory
+import com.hedvig.app.di.viewmodel.ViewModelFactory
 import com.hedvig.app.feature.claims.service.ClaimsTracker
 import com.hedvig.app.feature.claims.ui.ClaimsViewModel
 import com.hedvig.app.ui.fragment.RoundedBottomSheetDialogFragment
