+++ android/app/src/main/java/com/hedvig/app/feature/dashboard/ui/DashboardFragment.kt
@@ -13,22 +13,27 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
 import androidx.navigation.findNavController
-import com.hedvig.app.R
-import com.hedvig.app.di.ViewModelFactory
-import com.hedvig.app.feature.dashboard.service.DashboardTracker
 import com.hedvig.android.owldroid.graphql.DashboardQuery
 import com.hedvig.android.owldroid.type.DirectDebitStatus
 import com.hedvig.android.owldroid.type.InsuranceStatus
 import com.hedvig.android.owldroid.type.InsuranceType
-import com.hedvig.app.viewmodel.DirectDebitViewModel
+import com.hedvig.app.R
+import com.hedvig.app.di.viewmodel.ViewModelFactory
+import com.hedvig.app.feature.dashboard.service.DashboardTracker
 import com.hedvig.app.util.extensions.addViews
 import com.hedvig.app.util.extensions.compatDrawable
 import com.hedvig.app.util.extensions.observe
 import com.hedvig.app.util.extensions.setupLargeTitle
-import com.hedvig.app.util.extensions.view.*
+import com.hedvig.app.util.extensions.view.animateCollapse
+import com.hedvig.app.util.extensions.view.animateExpand
+import com.hedvig.app.util.extensions.view.remove
+import com.hedvig.app.util.extensions.view.setHapticClickListener
+import com.hedvig.app.util.extensions.view.show
+import com.hedvig.app.util.extensions.view.updatePadding
 import com.hedvig.app.util.interpolateTextKey
 import com.hedvig.app.util.isApartmentOwner
 import com.hedvig.app.util.isStudentInsurance
+import com.hedvig.app.viewmodel.DirectDebitViewModel
 import dagger.android.support.AndroidSupportInjection
 import io.reactivex.Flowable
 import io.reactivex.android.schedulers.AndroidSchedulers
