+++ android/app/src/main/java/com/hedvig/app/feature/offer/OfferChatOverlayFragment.kt
@@ -23,13 +23,13 @@ import com.facebook.react.ReactNativeHost
 import com.facebook.react.ReactRootView
 import com.facebook.react.common.LifecycleState
 import com.facebook.react.modules.core.DefaultHardwareBackBtnHandler
-import com.hedvig.app.di.ViewModelFactory
-import com.hedvig.app.util.extensions.localBroadcastManager
-import com.hedvig.app.NativeRoutingModule
-import com.hedvig.app.NativeRoutingModule.Companion.NAVIGATE_ROUTING_EXTRA_NAME_ACTION
-import com.hedvig.app.NativeRoutingModule.Companion.NAVIGATE_ROUTING_EXTRA_VALUE_LOGOUT_AND_RESTART_APPLICATION
 import com.hedvig.app.R
+import com.hedvig.app.di.viewmodel.ViewModelFactory
 import com.hedvig.app.feature.chat.ChatViewModel
+import com.hedvig.app.react.NativeRoutingModule
+import com.hedvig.app.react.NativeRoutingModule.Companion.NAVIGATE_ROUTING_EXTRA_NAME_ACTION
+import com.hedvig.app.react.NativeRoutingModule.Companion.NAVIGATE_ROUTING_EXTRA_VALUE_LOGOUT_AND_RESTART_APPLICATION
+import com.hedvig.app.util.extensions.localBroadcastManager
 import com.hedvig.app.util.showRestartDialog
 import dagger.android.support.AndroidSupportInjection
 import javax.inject.Inject
