+++ android/app/src/main/java/com/hedvig/app/feature/profile/ui/coinsured/CoinsuredFragment.kt
@@ -13,10 +13,14 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.findNavController
 import com.hedvig.app.R
-import com.hedvig.app.di.ViewModelFactory
+import com.hedvig.app.di.viewmodel.ViewModelFactory
 import com.hedvig.app.feature.profile.ui.ProfileViewModel
 import com.hedvig.app.util.CustomTypefaceSpan
-import com.hedvig.app.util.extensions.*
+import com.hedvig.app.util.extensions.compatColor
+import com.hedvig.app.util.extensions.compatFont
+import com.hedvig.app.util.extensions.compatSetTint
+import com.hedvig.app.util.extensions.concat
+import com.hedvig.app.util.extensions.setupLargeTitle
 import com.hedvig.app.util.extensions.view.remove
 import com.hedvig.app.util.extensions.view.show
 import dagger.android.support.AndroidSupportInjection
