+++ android/app/src/main/java/com/hedvig/app/feature/profile/ui/myhome/ChangeHomeInfoDialog.kt
@@ -10,7 +10,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.findNavController
 import com.hedvig.app.R
-import com.hedvig.app.di.ViewModelFactory
+import com.hedvig.app.di.viewmodel.ViewModelFactory
 import com.hedvig.app.feature.profile.ui.ProfileViewModel
 import com.hedvig.app.util.whenApiVersion
 import dagger.android.support.AndroidSupportInjection
