+++ android/app/src/main/java/com/hedvig/app/feature/profile/ui/payment/PaymentFragment.kt
@@ -13,16 +13,20 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.NavController
 import androidx.navigation.findNavController
-import com.hedvig.app.R
-import com.hedvig.app.di.ViewModelFactory
-import com.hedvig.app.feature.profile.ui.ProfileViewModel
 import com.hedvig.android.owldroid.type.DirectDebitStatus
-import com.hedvig.app.viewmodel.DirectDebitViewModel
+import com.hedvig.app.R
+import com.hedvig.app.di.viewmodel.ViewModelFactory
+import com.hedvig.app.feature.profile.ui.ProfileViewModel
 import com.hedvig.app.util.CustomTypefaceSpan
-import com.hedvig.app.util.extensions.*
+import com.hedvig.app.util.extensions.compatColor
+import com.hedvig.app.util.extensions.compatFont
+import com.hedvig.app.util.extensions.compatSetTint
+import com.hedvig.app.util.extensions.concat
+import com.hedvig.app.util.extensions.setupLargeTitle
 import com.hedvig.app.util.extensions.view.remove
 import com.hedvig.app.util.extensions.view.show
 import com.hedvig.app.util.interpolateTextKey
+import com.hedvig.app.viewmodel.DirectDebitViewModel
 import dagger.android.support.AndroidSupportInjection
 import kotlinx.android.synthetic.main.fragment_payment.*
 import kotlinx.android.synthetic.main.loading_spinner.*
