+++ android/app/src/main/java/com/hedvig/app/feature/profile/ui/payment/TrustlyFragment.kt
@@ -15,14 +15,14 @@ import android.webkit.WebView
 import android.webkit.WebViewClient
 import androidx.navigation.findNavController
 import com.hedvig.app.R
-import com.hedvig.app.di.ViewModelFactory
+import com.hedvig.app.di.viewmodel.ViewModelFactory
 import com.hedvig.app.feature.profile.ui.ProfileViewModel
-import com.hedvig.app.viewmodel.DirectDebitViewModel
 import com.hedvig.app.util.extensions.compatColor
 import com.hedvig.app.util.extensions.compatSetTint
 import com.hedvig.app.util.extensions.observe
 import com.hedvig.app.util.extensions.view.remove
 import com.hedvig.app.util.extensions.view.show
+import com.hedvig.app.viewmodel.DirectDebitViewModel
 import dagger.android.support.AndroidSupportInjection
 import kotlinx.android.synthetic.main.fragment_trustly.*
 import kotlinx.android.synthetic.main.loading_spinner.*
