+++ android/app/src/main/java/com/hedvig/app/feature/profile/ui/referral/ReferralFragment.kt
@@ -14,10 +14,14 @@ import android.view.ViewGroup
 import android.view.animation.OvershootInterpolator
 import androidx.navigation.findNavController
 import com.hedvig.app.R
-import com.hedvig.app.di.ViewModelFactory
+import com.hedvig.app.di.viewmodel.ViewModelFactory
 import com.hedvig.app.feature.profile.service.ProfileTracker
 import com.hedvig.app.feature.profile.ui.ProfileViewModel
-import com.hedvig.app.util.extensions.*
+import com.hedvig.app.util.extensions.compatColor
+import com.hedvig.app.util.extensions.compatDrawable
+import com.hedvig.app.util.extensions.compatSetTint
+import com.hedvig.app.util.extensions.observe
+import com.hedvig.app.util.extensions.setupLargeTitle
 import com.hedvig.app.util.extensions.view.increaseTouchableArea
 import com.hedvig.app.util.extensions.view.show
 import com.hedvig.app.util.interpolateTextKey
