+++ android/app/src/main/java/com/hedvig/app/util/react/AsyncStorageNativeImpl.kt
@@ -1,10 +1,9 @@
-package com.hedvig.app
+package com.hedvig.app.util.react
 
 import android.content.ContentValues
 import android.content.Context
 import android.database.Cursor
 import com.facebook.react.modules.storage.ReactDatabaseSupplier
-import com.hedvig.app.util.react.AsyncStorageNative
 import timber.log.Timber
 import javax.inject.Inject
